% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MFDI.R
\name{calc_MFDI}
\alias{calc_MFDI}
\title{Calculate Moorland Fire Danger Index}
\usage{
calc_MFDI(Rain, Temp, RH, Wind, Age, Time, Daily = FALSE)
}
\arguments{
\item{Rain}{Rainfall amount (mm)}

\item{Temp}{Temperature (degrees C)}

\item{RH}{Relative humidity (RH)}

\item{Wind}{Wind speed}

\item{Age}{Age of fuel (time since fire)}

\item{Time}{Vector of time values in minutes or days}

\item{Daily}{If time data is daily set to TRUE, otherwise if it is in minutes set to FALSE}
}
\value{
Moorland FDI
}
\description{
Calculate Moorland Fire Danger Index
}
\examples{
calc_MFDI(Rain=c(0,20,10,4),
Temp=c(12,13,8,9),
RH=c(55,87,90,67),
Wind=c(5,8,9,9),
Age=c(25,25,25,25),
Time=c(1,2,3,4),
Daily=TRUE)
}
