% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SDI.R
\name{calc_SDI}
\alias{calc_SDI}
\title{Calculate Mount's Soil Dryness Index (SDI)}
\usage{
calc_SDI(Rain, Temperature)
}
\arguments{
\item{Rain}{Daily precipitation in mm}

\item{Temperature}{Daily maximum temperature in degrees C}
}
\value{
Mount's SDI in mm
}
\description{
Calculate Mount's Soil Dryness Index (SDI)
}
\examples{
calc_SDI(c(2,5,0,0),c(26,17,18,20))
}
