% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_view_factor.R
\name{calc_view_factor}
\alias{calc_view_factor}
\title{Calculate View Factor}
\usage{
calc_view_factor(FlameAngle, dist, site_slope, Lf, h, Wf)
}
\arguments{
\item{FlameAngle}{Distance (metres)}

\item{dist}{Distance (metres)}

\item{site_slope}{Slope of site (degrees)}

\item{Lf}{Flame length}

\item{h}{Elevation of receiver (m)}

\item{Wf}{Flame width}
}
\value{
View Factor
}
\description{
Calculate View Factor
}
\examples{

calc_view_factor(20,15,5,5,2,100)
}
