% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_DEM.R
\docType{data}
\name{COH_DEM}
\alias{COH_DEM}
\title{Example digital elevation model (DEM) for landscape fire risk calculation}
\source{
Tas Government
}
\usage{
COH_DEM
}
\description{
DEM for region of Hobart, Tasmania
}
\author{
Tas Government
}
