% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Fire.R
\docType{data}
\name{COH_Fire}
\alias{COH_Fire}
\title{Fire History polygon example data}
\source{
Tas LIST
}
\usage{
COH_Fire
}
\description{
Polygons of fire boundaries for an area west of Hobart, Tasmania
derived from the Tasmanian Fire History dataset, with attribute for season of burn
}
\author{
Tas LIST
}
