% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Fuel.R
\docType{data}
\name{COH_Fuel}
\alias{COH_Fuel}
\title{Example fuels data, Tasmanian vegetation types}
\source{
Tas Government
}
\usage{
COH_Fuel
}
\description{
Data frame with fuel types, formations, fine fuel loads and heights.  SFFL is surface fine fuel load (t/ha), OFFL = overall fuel load (t/ha) and HGT is elevated fuel height in metres.
}
\author{
Tas Government
}
