% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Sites.R
\docType{data}
\name{COH_Sites}
\alias{COH_Sites}
\title{Example input point data for landscape risk calculation}
\source{
Tas Government
}
\usage{
COH_Sites
}
\description{
sf point object with 5 locations in Hobart, Tasmania
}
\author{
Tas Government
}
