% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Veg.R
\docType{data}
\name{COH_Veg}
\alias{COH_Veg}
\title{Vegetation polygon data}
\source{
Tas LIST / DPIPWE 2020
}
\usage{
COH_Veg
}
\description{
Demonstration vegetation polygon data with vegetation code, group, forest type and ID
}
\author{
Tas LIST / DPIPWE 2020
}
