% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Veg_LUT.R
\docType{data}
\name{COH_Veg_LUT}
\alias{COH_Veg_LUT}
\title{Vegetation fire attribute look-up table}
\source{
Tas DPIPWE
}
\usage{
COH_Veg_LUT
}
\description{
Look-up table linking vegetation codes to maximum and minimum fire intervals
}
\author{
Tas DPIPWE
}
