% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COH_Veg_Risk.R
\docType{data}
\name{COH_Veg_Risk}
\alias{COH_Veg_Risk}
\title{Example vegetation map for landscape fire risk calculation}
\source{
Tas Government
}
\usage{
COH_Veg_Risk
}
\description{
Vegetation polygon layer with vegetation class text
}
\author{
Tas Government
}
