% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_vectors.R
\name{compass_vectors}
\alias{compass_vectors}
\title{Construct 8 line vectors in cardinal directions from a point over a distance}
\usage{
compass_vectors(point, distance)
}
\arguments{
\item{point}{sf object with a point}

\item{distance}{Distance in metres}
}
\value{
sf object with line vectors
}
\description{
Construct 8 line vectors in cardinal directions from a point over a distance
}
\examples{

 compass_vectors(sf::st_as_sf(data.frame(x=521310,y=5244614),coords=c("x","y"),crs=28355),50)
}
