% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayintegral.R
\name{dayintegral}
\alias{dayintegral}
\title{Calculate number of hours of sunlight for a given day}
\usage{
dayintegral(dayofyear, latitude, precision = 64)
}
\arguments{
\item{dayofyear}{Julian day of year}

\item{latitude}{Latitude of location}

\item{precision}{Precision of integral calculation (default 64)}
}
\value{
Hours of sunlight
}
\description{
Calculate number of hours of sunlight for a given day
}
\examples{
dayintegral(20,-43)
}
