% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylengths.R
\name{daylengths}
\alias{daylengths}
\title{Calculate a vector of day lengths for a given latitude, by julian day}
\usage{
daylengths(latitude, precision = 64)
}
\arguments{
\item{latitude}{Latitude of location}

\item{precision}{Optional, precision of day integral (default value 64)}
}
\value{
Vector of day lengths in hours
}
\description{
Calculate a vector of day lengths for a given latitude, by julian day
}
\examples{
daylengths(-43)
}
