% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineMaxViewFactor.R
\name{determineMaxViewFactor}
\alias{determineMaxViewFactor}
\title{Determine Maximum View Factor}
\usage{
determineMaxViewFactor(dist, Lf, site_slope, h, Wf)
}
\arguments{
\item{dist}{Distance (metres)}

\item{Lf}{Flame length}

\item{site_slope}{Slope of site (degrees)}

\item{h}{Elevation of receiver (m)}

\item{Wf}{Flame width}
}
\value{
MaxViewFactor
}
\description{
Determine Maximum View Factor
}
\examples{

determineMaxViewFactor(20,5,5,2,100)
}
