% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ember_dist.R
\name{ember_dist}
\alias{ember_dist}
\title{Modelled maximum 5 ember/m-2 ember distance}
\usage{
ember_dist(FFDI, FL)
}
\arguments{
\item{FFDI}{McArthur Forest Fire Danger Index}

\item{FL}{Fuel Load in t/ha}
}
\value{
Distance in metres of outer ember transport distance
}
\description{
Modelled maximum 5 ember/m-2 ember distance
}
\examples{

ember_dist(45,15.5)
}
