% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBW_calc.R
\name{FBW_calc}
\alias{FBW_calc}
\title{Fuel Break Width Calculator}
\usage{
FBW_calc(
  fb_class,
  vegetation_community,
  slope_type,
  effective_slope,
  max_fire_run_distance,
  datdf,
  dist = 0
)
}
\arguments{
\item{fb_class}{Fuel break class; string class1 = Managed fuel, class2 = Protective break only}

\item{vegetation_community}{Vegetation community from the Vegetation_Community field of fuel load data frame}

\item{slope_type}{Position of property; string upslope, downslope or none}

\item{effective_slope}{Slope in degrees}

\item{max_fire_run_distance}{Maximum fire run distance in metres}

\item{datdf}{Data frame with Vegetation_Community, SFFL (surface) and OFFL (overall) fine fuel loads, understory height in HGT, and Vegetation_Category ("Forest, Rainforest and Woodland", "Shrub, Scrub and Heath","Grassland" or "Tussock Moorland") to select fire model}

\item{dist}{Distance field, unsure purpose}
}
\value{
Maximum fuel break width in m
}
\description{
Fuel Break Width Calculator
}
\examples{
\dontrun{

FBW_calc("class1","DOB","downslope",5,340,fuel_data)
}
}
