% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffdi.R
\name{ffdi}
\alias{ffdi}
\title{Calculator for single Forest Fire Danger Index calculation}
\usage{
ffdi(Temperature, Humid, Wind, DF)
}
\arguments{
\item{Temperature}{Temperature in C}

\item{Humid}{Relative Humidity in \%}

\item{Wind}{Wind Speed in km/h}

\item{DF}{Drought factor (1-10)}
}
\value{
McArthur FFDI value
}
\description{
Calculator for single Forest Fire Danger Index calculation
}
\examples{

ffdi(18,28,28,5.4)
}
