% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_behav.R
\name{forest_behav}
\alias{forest_behav}
\title{Forest fire behaviour parameters from FFDI and Fuel Load}
\usage{
forest_behav(FFDI, FL, GS = 0)
}
\arguments{
\item{FFDI}{McArthur Forest Fire Danger Index}

\item{FL}{Fuel Load in t/ha}

\item{GS}{Ground slope in degrees}
}
\value{
Data frame containing columns for flame height (m), spotting distance (km), rate of spread (km/h) and fireline intensity (kW/m)
}
\description{
Forest fire behaviour parameters from FFDI and Fuel Load
}
\examples{

forest_behav(56,12.5,5)
}
