% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_max_run.R
\name{generate_max_run}
\alias{generate_max_run}
\title{Precalculate max fire run distance}
\usage{
generate_max_run(
  locations,
  veg,
  urb_types = c("Other natural environments",
    "Agricultural, urban and exotic vegetation"),
  urb_set = "Agricultural, urban and exotic vegetation",
  id_field = "RespondentID"
)
}
\arguments{
\item{locations}{sf point object containing an ID field}

\item{veg}{Vegetation polygon layer with vegetation type in VEG_GROUP field}

\item{urb_types}{List of vegetation types to treat as urban/no fire risk}

\item{urb_set}{Vegetation type string to associate with urban in output}

\item{id_field}{Name of field contianing ID}
}
\value{
sf point object with landscape fields added, ready for risk calculation
}
\description{
Precalculate max fire run distance
}
\examples{
\dontrun{
output <- generate_max_run(data_points,veg,id_field="ID")
}
}
