% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_property_fields.R
\name{generate_property_fields}
\alias{generate_property_fields}
\title{Precalculate landscape fields reqired for landscape risk calculations}
\usage{
generate_property_fields(
  locations,
  dem,
  veg,
  urb_types = c("Other natural environments",
    "Agricultural, urban and exotic vegetation"),
  urb_set = "Agricultural, urban and exotic vegetation",
  id_field = "RespondentID"
)
}
\arguments{
\item{locations}{sf point object containing an ID field}

\item{dem}{Raster object with digtial elevation model of region}

\item{veg}{Vegetation polygon layer with vegetation type in VEG_GROUP field}

\item{urb_types}{List of vegetation types to treat as urban/no fire risk}

\item{urb_set}{Vegetation type string to associate with urban in output}

\item{id_field}{Name of field contianing ID}
}
\value{
sf point object with landscape fields added, ready for risk calculation
}
\description{
Precalculate landscape fields reqired for landscape risk calculations
}
\examples{
\dontrun{
output <- generate_property_fields(data_points,dem,veg,id_field="ID")
}
}
