% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ET.R
\name{get_ET}
\alias{get_ET}
\title{Get Evapotranspiration value for given Temperature and Soil Dryness Index}
\usage{
get_ET(Temp, SDI)
}
\arguments{
\item{Temp}{Temperature in Degrees C}

\item{SDI}{Mount's Soil Dryness Index}
}
\value{
Evapotranspiration value
}
\description{
Get Evapotranspiration value for given Temperature and Soil Dryness Index
}
\examples{
get_ET(15,221)
}
