% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moor_FM.R
\name{get_moor_FM}
\alias{get_moor_FM}
\title{Calculates Moorland FDI Fuel Moisture}
\usage{
get_moor_FM(Temp, Humid, RainHours, RainAmount)
}
\arguments{
\item{Temp}{Temperature in degrees C}

\item{Humid}{Humidity percentage}

\item{RainHours}{Number of hours over which rainfall fell}

\item{RainAmount}{Amount of rainfall (mm)}
}
\value{
Moorland Fuel Moisture
}
\description{
Calculates Moorland FDI Fuel Moisture
}
\examples{
get_moor_FM(12,87,24,22)
}
