% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moor_HF.R
\name{get_moor_HF}
\alias{get_moor_HF}
\title{Calculate moorland FDI humidity factor}
\usage{
get_moor_HF(Temp, Humid)
}
\arguments{
\item{Temp}{Temperature in degrees C}

\item{Humid}{Humidity percentage}
}
\value{
Moorland FDI Humidity Factor
}
\description{
Calculate moorland FDI humidity factor
}
\examples{
get_moor_HF(12,87)
}
