% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moor_RF.R
\name{get_moor_RF}
\alias{get_moor_RF}
\title{Calculate Moorland FDI Rainfall Factor}
\usage{
get_moor_RF(RainHours, RainAmount)
}
\arguments{
\item{RainHours}{Number of hours over which rainfall fell}

\item{RainAmount}{Amount of rain in mm}
}
\value{
Moorland FDI Rainfall Factor
}
\description{
Calculate Moorland FDI Rainfall Factor
}
\examples{
get_moor_RF(24,20)
}
