% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moor_RS.R
\name{get_moor_RS}
\alias{get_moor_RS}
\title{Calculate Moorland FDI rate of spread}
\usage{
get_moor_RS(FM, Wind, Age)
}
\arguments{
\item{FM}{Fuel moisture as calculated by get_moor_FM}

\item{Wind}{Wind speed}

\item{Age}{Age of fuel (time since fire)}
}
\value{
Rate of Spread
}
\description{
Calculate Moorland FDI rate of spread
}
\examples{
get_moor_RS(18,10,25)
}
