% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourwatt.R
\name{hourwatt}
\alias{hourwatt}
\title{Calculate radiation in W/m2 for a given latitude, hour of day and day of year}
\usage{
hourwatt(dayofyear, latitude, hour)
}
\arguments{
\item{dayofyear}{Julian Day of year}

\item{latitude}{Latitude of location}

\item{hour}{Hour after midnight}
}
\value{
Radiation (W/m2)
}
\description{
Calculate radiation in W/m2 for a given latitude, hour of day and day of year
}
\examples{
hourwatt(20,-43,17)
}
