% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_vector.R
\name{line_vector}
\alias{line_vector}
\title{Construct line vector from a point over a given distance}
\usage{
line_vector(point, angle, distance)
}
\arguments{
\item{point}{sf object with a point}

\item{angle}{Compass angle in degrees}

\item{distance}{Distance in metres}
}
\value{
sf LineString object of vector
}
\description{
Construct line vector from a point over a given distance
}
\examples{
\dontrun{

line_vector(point,15,50)
}
}
