% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_vector.R
\name{nw_vector}
\alias{nw_vector}
\title{Generate sf vector line in a north-west direction from point}
\usage{
nw_vector(point, dist = 350)
}
\arguments{
\item{point}{sf point object}

\item{dist}{Distance in meters}
}
\value{
Data frame with slope, r2 value and angle in degrees
}
\description{
Generate sf vector line in a north-west direction from point
}
\examples{
nw_vector(sf::st_as_sf(data.frame(x=521310,y=5244614),coords=c("x","y"),crs=28355),350)
}
