% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq.R
\name{pq}
\alias{pq}
\title{Print non-quiet information.}
\usage{
pq(text, quiet)
}
\arguments{
\item{text}{Text to be printed}

\item{quiet}{TRUE/FALSE}
}
\description{
Print non-quiet information.
}
\examples{
pq("Hello",TRUE)
}
\keyword{internal}
