% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFire.R
\name{processFire}
\alias{processFire}
\title{Process input fire history vector data to produce yearly history rasters}
\usage{
processFire(
  outdir,
  fire_history,
  season_field,
  start_year = NULL,
  end_year = NULL,
  quiet = TRUE
)
}
\arguments{
\item{outdir}{Directory to output fire history, must contains files produced by createDomain}

\item{fire_history}{SpatVect or SimpleFeatures object contianing the fire history polygons. Must contain a field with fire date as a four-digit year.}

\item{season_field}{Name of field in SpatVect containing year/season of fire}

\item{start_year}{Optional, start year for history calculation. Years before this will be excluded.}

\item{end_year}{Optional end year for history production. Years after this will be excluded. If this year is beyond the end year in the history, empty future projection rasters will be produced.}

\item{quiet}{Print progress information}
}
\value{
Path to outdir for piping.
}
\description{
Process input fire history vector data to produce yearly history rasters
}
\examples{
\dontrun{
data("COH_Fire")
fire_history <- vect(COH_Fire)
processFire("output",fire_history=fire_data,season_field="Year")
}
}
