% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAD_calc.r
\name{RAD_calc}
\alias{RAD_calc}
\title{Radiant Heat Calculator}
\usage{
RAD_calc(
  FFDI = 24,
  GFDI = 24,
  V = 26,
  Ta = 31,
  Ws = 20,
  Wo = 30,
  HGT = -1,
  fuel_type = "Forest",
  slope_type = "upslope",
  effective_slope = 5,
  max_fire_run_distance = "< 300m",
  dist = 50
)
}
\arguments{
\item{FFDI}{Forest Fire Danger Index}

\item{GFDI}{Grassland Fire Danger Index}

\item{V}{10m Wind Speed (kmh)}

\item{Ta}{Ambient Temperature (C)}

\item{Ws}{Surface Fuel Load (t/ha)}

\item{Wo}{Overall Fuel load (t/ha)}

\item{HGT}{Understory height (m, for heathland, grassland or shrub, all others set to -1)}

\item{fuel_type}{(string, Forest/Shrub/Grass/Moorland)}

\item{slope_type}{Position of property; string upslope, downslope or none}

\item{effective_slope}{Slope in degrees}

\item{max_fire_run_distance}{Maximum fire run distance in metres}

\item{dist}{Distance of house from fireline (m)}
}
\value{
Radiant heat flux in kW/m2
}
\description{
Radiant Heat Calculator
}
\examples{
\dontrun{

RAD_calc()
}
}
