% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\docType{data}
\name{ROI}
\alias{ROI}
\title{Region of Interest}
\source{
Grant Williamson 2022
}
\usage{
ROI
}
\description{
Demonstration Region of Interest polygon (ROI) to carry out analysis within.
}
\author{
Grant Williamson 2022
}
