% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_to_MFI.R
\name{RS_to_MFI}
\alias{RS_to_MFI}
\title{Convert Rate of Spread to Moorland FDI}
\usage{
RS_to_MFI(RS)
}
\arguments{
\item{RS}{Rate of spread}
}
\value{
Moorland FDI
}
\description{
Convert Rate of Spread to Moorland FDI
}
\examples{
RS_to_MFI(4.5)
}
