% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_poly.R
\name{shift_poly}
\alias{shift_poly}
\title{Extend and shift a polygon by a given angle and distance}
\usage{
shift_poly(poly, dis = 50, base_ang = 135, var_ang = 6.5)
}
\arguments{
\item{poly}{A polygon sf object}

\item{dis}{Distance to extend in metres}

\item{base_ang}{Angle in which to extend polygon}

\item{var_ang}{Lateral angle range in which to shuffle polygon}
}
\value{
A sf polygon object
}
\description{
Extend and shift a polygon by a given angle and distance
}
\examples{
\dontrun{

shift_poly(vegetation,50,135,6.5)
}
}
