% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_vector.R
\name{slope_vector}
\alias{slope_vector}
\title{Calculate slope based on regression over a given distance and angle across a DEM}
\usage{
slope_vector(point, distance, angle, dem)
}
\arguments{
\item{point}{sf object with a point}

\item{distance}{Distance in metres}

\item{angle}{Compass vector angle}

\item{dem}{RASTER dem object}
}
\value{
Slope
}
\description{
Calculate slope based on regression over a given distance and angle across a DEM
}
\examples{

 slope_vector(sf::st_as_sf(data.frame(x=526000,y=5248614),coords=c("x","y"),crs=28355),150,37,dem)
}
