% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslp.R
\name{sslp}
\alias{sslp}
\title{Regress slope from vector of elevations}
\usage{
sslp(y, dist = 150)
}
\arguments{
\item{y}{Vector of elevations spread over the distance}

\item{dist}{Distance in meters}
}
\value{
Data frame with slope, r2 value and angle in degrees
}
\description{
Regress slope from vector of elevations
}
\examples{
sslp(c(20,30,45,40,80,91))
}
