% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunaltitude.R
\name{sunaltitude}
\alias{sunaltitude}
\title{Calculate sun altitude for given latitude, day and time}
\usage{
sunaltitude(latitude, dayofyear, hour)
}
\arguments{
\item{latitude}{Latitude of location}

\item{dayofyear}{Julian day of year}

\item{hour}{Time of day (hours from midnight)}
}
\value{
Sunaltitude in degrees
}
\description{
Calculate sun altitude for given latitude, day and time
}
\examples{
sunaltitude(-43,20,15)
}
