% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomain.R
\name{createDomain}
\alias{createDomain}
\title{Create template and region of interest (ROI) rasters in an output directory based on input vector ROI, resolution and projection.}
\usage{
createDomain(domain, res, proj, outdir)
}
\arguments{
\item{domain}{SpatVect or SimpleFeatures object contianing the domain of the analysis.  If this contains multiple features, they will be aggregated.}

\item{res}{Resolution, in project units, of the output template and mask rasters}

\item{proj}{Projection (CRS) object for the template and mask rasters to be produced}

\item{outdir}{Output directory to write the rasters}
}
\value{
template
}
\description{
Create template and region of interest (ROI) rasters in an output directory based on input vector ROI, resolution and projection.
}
\examples{
createDomain(myVect, 30, "EPSG:3308","./output")
}
